Made features on user-screen.html work
Did some corrections and adjustments for user-screen-html and user-screen.js. Now at least name and eMail can be entered in text boxes and the Add button shows some functionality. A title has been but it is misplaced and the required data are not being fetched. Issues how to merge user- with users-screen remain.